Clippy fixes

+++ diesel_derives/src/diesel_for_each_tuple.rs
@@ -17,7 +17,7 @@ pub(crate) fn expand(input: Ident) -> TokenStream {
             let t = Ident::new(&format!("T{}", i), call_side);
             let st = Ident::new(&format!("ST{}", i), call_side);
             let tt = Ident::new(&format!("TT{}", i), call_side);
-            let i = syn::Index::from(i as usize);
+            let i = syn::Index::from(i);
             quote!((#i) -> #t, #st, #tt,)
         })
         .collect::<Vec<_>>();
